chore: Add Tron HTTP client for contract retrieval and testing
This commit adds a new Tron HTTP client in the `tron_http.go` file. The client provides a method `GetContract` to retrieve contract information from the Tron blockchain. It also includes a corresponding test in the `tron_http_test.go` file to ensure the functionality of the client.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.