Refactor Tron assembler functions to include transaction timestamp
This commit modifies the Tron assembler functions `FromTronTransferContract`, `FromTronTriggerSmartContract`, and `FromTronContractToTransaction` to include the transaction timestamp in the created `Transaction` objects. The timestamp is extracted from the `TransactionExtention` object and converted to a `time.Time` value using the `time.Unix` function. This change ensures that the `CreatedAt` field of the `Transaction` objects accurately reflects the timestamp of the corresponding transactions.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.