refactor: Update EthereumTransactionsSerialize test to use interface{} for JSON marshaling
This commit updates the `TestEthereumTransactionsSerialize` function in the `eth_serialization_test.go` file to use the `interface{}` type for JSON marshaling. Instead of directly marshaling the `txs` variable, it assigns it to an `interface{}` variable `o` and marshals `o`. This change improves flexibility and allows for easier modification of the marshaling process in the future.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.