refactor: Update EthereumBlockChainService to use channel for transaction updates
This commit updates the `EthereumBlockChainService` struct in the `eth.go` file to use a channel for sending transaction updates. It adds a `Channel` field of type `chan interface{}` to the struct and modifies the `GetBlock` function to send the transactions to the channel. This change enables real-time transaction updates and improves the responsiveness of the monitoring system.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.