test: Add unit test for TestCommandParameter
This commit adds a unit test for the `TestCommandParameter` function in the `common_test.go` file. The test case verifies that the subcommand is executed and the logging flag is set to true when the `subcommand` command is executed with the `--logging` flag. This addition ensures the correctness of the command parameter functionality and helps maintain the integrity of the command package.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.