refactor: Update main.go and commands package
This commit refactors the `main.go` file and the `commands` package. It introduces two new functions, `NewMonitorCobraCommand` and `NewRootCommand`, to create the monitor and root commands respectively. It also adds the `RunCommand` function to execute the root command. Additionally, it moves the `SetupLogger` function from `commands/initialization.go` to `main.go` and renames it to `setupLogger`. This refactoring improves code organization and separation of concerns.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.