refactor: Update EthereumBlockChainService to accept additional parameters
This commit updates the `EthereumBlockChainService` struct in the `eth.go` file to accept additional parameters. It modifies the `NewEthereumBlockChainService` function to include a new parameter `options` of type `interface{}`. This change allows for more flexibility and customization when creating an instance of the Ethereum block chain service.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.