refactor: Update README.md with badges and remove extra whitespace
This commit adds badges to the README.md file, including GoDoc, Forum, Twitter, and License badges. It also removes extra whitespace at the end of the file. These changes improve the documentation and provide additional resources for users to engage with the project.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.