chore: Update Go dependencies
This commit updates the Go dependencies in the `go.mod` and `go.sum` files. It adds the `golang.org/x/exp` package as an indirect dependency with version `v0.0.0-20231110203233-9a3e6036ecaa`. Additionally, it updates the `github.com/ethereum/go-ethereum` package to version `v1.14.5`. These dependency updates ensure that the project is using the latest stable versions of the required packages.

Note: This commit message is based on the provided code changes and does not take into account the recent repository commits.